+++ R/ggpiestats.R
@@ -110,7 +110,7 @@ ggpiestats <- function(
   if (".counts" %in% names(data)) data %<>% tidyr::uncount(weights = .counts)
 
   # x and y need to be a factor; also drop the unused levels of the factors
-  data %<>% mutate(across(.cols = everything(), .fns = ~ droplevels(as.factor(.x))))
+  data %<>% mutate(across(.cols = everything(), .fns = ~ as.factor(.x)))
   x_levels <- nlevels(pull(data, {{ x }}))
   y_levels <- ifelse(test == "one.way", 0L, nlevels(pull(data, {{ y }})))